+++ rentplace/src/main/java/kattsyn/dev/rentplace/dtos/properties/PropertyCreateEditDTO.java
@@ -1,4 +1,4 @@
-package kattsyn.dev.rentplace.dtos;
+package kattsyn.dev.rentplace.dtos.properties;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
